requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
jsonpath==0.82.2
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
wsproto==1.2.0
zipp==3.15.0
